safari-534.53-branch/LayoutTests/fast/js/regexp-range-bound-ffff-expected.txt

Test for rdar:/68455379, a case-insensitive regex containing a character class containing a range with an upper bound of ￿ can lead to an infinite-loop.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS ("A".match(/[-￿]/i) == "A") is true
PASS successfullyParsed is true

TEST COMPLETE
